[ADD] Module  help_online
This module allows the creation of an online help available from the
lists and forms in Odoo.

When loading a view, the module generates a button allowing access to
an help page for the related model if the page exists and the user is
member of the group 'Help reader'. If the page doesn't exist and the
user is member of the group 'Help writer', the module generate a button
allowing the creation an help page.

The help pages are created and managed via the website Module.