Update __openerp__.py
In response to Daniel's comments I have modified __openerp__.py , I apologise for a lack of adherence to the guidelines. I am new to using GITHUB and contributing to OCA and so I am trying to get a handle on all of these details. Thank you for your advice and patience.