[FIX][web_export_view] Remove monetary formatting
Monetary fields were hard to work with in Excel after exporting because having the currency symbol made Excel treat those cells as text.

This patch removes those symbols and thus allows Excel to treat those values as numbers, allowing easier management.

It works no matter the symbol is before or after the value, so in any case it fixes #558.