[FIX] web_widget_domain_v11: Allow to change operator
Previously, a Bootstrap dropdown was being used to choose the domain operator.

It was not working due to a problem that doesn't allow to have nested dropdowns in Odoo 10.0 search views, but possibly yes in the intermediate version where this addon was backported.

To get this working, the BS dropdown is replaced by a standard select element, so no more conflicts.