[fix] web_translate_dialog: correct context split
Fixes the display of languages other than the user language  in the translation dialog.
split_context checks if we provide more arguments than read has and then uses of the additional one as context.
Since read takes the argument load additionally to the ones already given, we need to provide it as well to ensure the context is split.