web_ckeditor4: fix no re-init on form discard
When you hit "cancel" button or when you navigate away from the form,
for instance by clicking on the breadcrumb or on  "edit translations",
we have to remove the CKEditor widget.
It happens the same if you click on a main menu item
(eg: you are on sale order create form and you click on
Sales -> Quotations to get back to the tree view).

BUT then if you hit "create" Odoo's form machinery is not initializing
the widget anymore, which really sounds inconsistent.
If the widget is not initialized again it means that if CKEditor
got destroyed there's no way to re-init again.

Here we make sure that on create (no id on datarecord)
if the editor is not initialized yet we force it.