[IMP] web_m2x_options: new option 'delay' on Many2one
This option allows to delay the 'name_search' RPC request done when the
user write in Many2one fields.

By default in Odoo the request is performed after 200ms, which can be too fast,
generating a lot of RPC queries stressing uselessly the server.
Increasing this value (for instance to 800ms) allows to get better performance
as the server is less stressed.

As other options, it can be set globally (`ir.config_parameter`) or at
the field level.