web_translate_dialog: Display only fields displayed in the view
parent.renderer.state.fields includes all the fields in the model
whereas parent.renderer.state.getFieldNames() includes all the fields
displayed in the view.
This commit filters out fields that are not displayed in the view
to avoid the TypeError "undefined is not an object (evaluating '$('[name='+field_name+']')[0].maxLength')"
because translatable_fields did include fields that are not in the view.