[12.0][IMP] web_m2x_options: Add search_more_always option
Adds a new option search_more_always that when set to True
in system parameters, it ALWAYS shows "search more..." option even if
there are no records or records number < limit.

This is useful especially when the referenced table in many2one has
active field and you want to select a record that is archived.