[FIX] web_widget_numeric_step: Remove focus to avoid annoying flickering effect
The focus introduced on the 16.0 migration is causing a flickering effect
that is a bit annoying and does not make complete sense, because if you
click more than once the cursor is positioned at the end of the text,
which makes it difficult to edit the input.

The counterpart of these changes is that to edit the input using the
keyboard we will have to click on the field. Although this is how the
module has been operating since its inception.