[FIX] web_responsive: prevent error in tests when accessing an undefined variable
For an unknown reason, when running the tests, the session does not have all variables filled, so apps_menu is undefined.
This causes an error in tours of other modules: "Cannot read properties of undefined (reading 'theme')".

The curious thing is that if you run the tests of this module alone, they pass.
But when running tests together with another module, tests from the other module fail.
For example, when running tests with the voip_oca module.